% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{amg}
\alias{amg}
\title{Run AMG preconditioner on vector}
\usage{
amg(P, b)
}
\arguments{
\item{P}{an amgcl preconditioner object}

\item{b}{vector to run through preconditioner}
}
\value{
the product P*b
}
\description{
Unlike psolve, which solves a linear system (up to a certain tolerance), this function just runs the AMG preconditioner on a vector.
You can think of the preconditioner as a fast approximate matrix inverse. In a few cases it is an exact inverse, for example in the case of a Laplacian on a lattice.
}
\examples{

}
\author{
Simon Barthelme
}
