% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{asolve}
\alias{asolve}
\title{Solve linear system iteratively using  algebraic multigrid preconditioning}
\usage{
asolve(A, b, guess, null, tol = 1e-05, maxiter = 100,
  coarsen = "smoothed_aggregation", relax = "spai0", max_level = 4,
  solver = "gmres")
}
\arguments{
\item{A}{sparse, square matrix}

\item{b}{right-hand side}

\item{guess}{(optional) initial value for the iteration}

\item{null}{(optional) null space of A, for singular systems}

\item{tol}{tolerance}

\item{maxiter}{maximum number of iterations}

\item{coarsen}{coarsening method to use (one of the following: ruge_stuben,aggregation,smoothed_aggregation,smoothed_aggr_emin)}

\item{relax}{relaxation to use (one of the following: gauss_seidel,multicolor_gauss_seidel,ilu0,parallel_ilu0,iluk,ilut,damped_jacobi,spai0,spai1,chebyshev)}
}
\value{
a vector x such that Ax=b (or as far as the iteration would allow)
}
\description{
Interface to Denis Demidov's amgcl library
}
\examples{
A <- sparseMatrix(1:2,1:2,x=rnorm(2))
b <- rnorm(2)
asolve(A,b)
#Compare to: 
solve(A,b)
#Use on singular matrices:
u <- sparseVector(rnorm(3),1:3,3)
A <- u\%*\%t(u)
#A has a null space:
null <- svd(A,3)$v[,2:3]
A\%*\%null
b <- A\%*\%rnorm(3)
asolve(A,b,null=null,max_level=1)
}
\author{
Simon Barthelme
}
