% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{precond}
\alias{precond}
\alias{psolve}
\title{Solve a linear system iteratively using  algebraic multigrid preconditioning}
\usage{
precond(A, tol = 1e-05, maxiter = 100, coarsen = "smoothed_aggregation",
  relax = "spai0", max_level = 4, solver = "gmres")

psolve(P, b, guess, A)
}
\arguments{
\item{A}{sparse, square matrix (from Matrix package)}

\item{tol}{tolerance (default 1e-5)}

\item{maxiter}{maximum number of iterations (default 100)}

\item{coarsen}{coarsening method to use (one of the following: ruge_stuben,aggregation,smoothed_aggregation,smoothed_aggr_emin)}

\item{relax}{relaxation to use (one of the following: gauss_seidel,ilu0,iluk,ilut,damped_jacobi,spai0,spai1,chebyshev)}

\item{max_level}{maximum number of levels in the multiscale hierarchy (default 4)}

\item{solver}{which iterative solver to use (cg, idrs, lgmres, gmres, bicgstab,bicgstabl). cg only works on SPD matrices, the rest allow non-symmetric matrices.}

\item{P}{a preconditioner (object of class "amgcl")}

\item{b}{right-hand side (numeric, real)}

\item{guess}{(optional) initial value for the iteration}
}
\value{
a vector x such that Ax=b (or as far as the iteration would allow), with attributes "iter" (number of iterations), and "error" (root mean squared residual).
}
\description{
Interface to Denis Demidov's amgcl library. There are two main functions, "precond" and "psolve". precond builds the preconditioner (an approximation to the system matrix) and sets up the solver. psolve runs the preconditioner on a specific right-hand side.
}
\details{
You can change the system matrix when running psolve (using the optional A argument ). The original preconditioner will be retained, which means you save the cost of setting up a new preconditioner, but the preconditioner is no longer optimally suited to the system. Use this only for small modifications to the system matrix.
}
\section{Functions}{
\itemize{
\item \code{psolve}: solve a preconditioned system
}}
\examples{
A <- sparseMatrix(1:2,1:2,x=rnorm(2))
b <- rnorm(2)
P <- precond(A)
psolve(P,b)
#Compare to: 
solve(A,b)


}
\author{
Simon Barthelme
}

